% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phyloseq_object.R
\name{create_phyloseq_object}
\alias{create_phyloseq_object}
\title{Create Phyloseq Object from CSV Files}
\usage{
create_phyloseq_object(abundance_file, metadata_file, output_file)
}
\arguments{
\item{abundance_file}{The path to the abundance CSV file.}

\item{metadata_file}{The path to the metadata CSV file.}

\item{output_file}{The name of the output Rdata file (e.g., 'aphy_sample.Rdata').}
}
\value{
The phyloseq object ps.ng.tax.
}
\description{
This function reads in abundance and metadata CSV files, creates a phyloseq object named ps.ng.tax, and saves it as an Rdata file.
}
