% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_degree.R
\name{extr_degree}
\alias{extr_degree}
\title{Extract Degree and Merge with Metadata and OTU Table}
\usage{
extr_degree(mypath, rdata_path, output_path)
}
\arguments{
\item{mypath}{A character string specifying the directory containing the .gml files.}

\item{rdata_path}{A character string specifying the directory containing the .Rdata files (phyloseq objects).}

\item{output_path}{A character string specifying the directory where output CSV files will be saved.}
}
\value{
No return value. The function writes output CSV files to the specified output directory.
}
\description{
This function reads .gml graph files from a specified directory, computes node degrees, extracts graph attributes, merges them with metadata and OTU tables from corresponding phyloseq objects, and writes the results to CSV files.
}
