% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_netattr.R
\name{extr_netattr}
\alias{extr_netattr}
\title{Extract Network Attributes}
\usage{
extr_netattr(
  mypath = "E:/diabetes0409/",
  phylo_list = c("aphy_diabetesmetross_resp.Rdata")
)
}
\arguments{
\item{graph}{The input graph from which attributes are extracted.}
}
\value{
A data frame containing the extracted attributes.
}
\description{
This function extracts network attributes from the provided graph.
}
