% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_edgelist.R
\name{Generate_edgelist}
\alias{Generate_edgelist}
\title{Generate Edge List from Microbial Data}
\usage{
Generate_edgelist(filename, bootstraps = 1000, p_value_threshold = 0.05)
}
\arguments{
\item{filename}{The path to the input file containing microbial count data.}

\item{bootstraps}{The number of bootstrap samples to use (default is 1000).}

\item{p_value_threshold}{The significance threshold for p-values (default is 0.05).}
}
\value{
A data frame containing the edge list with significant correlations.
}
\description{
This function reads microbial count data, performs centered log-ratio transformation,
computes SparCC correlations, performs bootstrapping to calculate p-values,
and generates an edge list of significant correlations.
}
