% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_samplespecificabundance.R
\name{generate_samplespecificabundance}
\alias{generate_samplespecificabundance}
\title{Generate Sample-Specific Abundance Data}
\usage{
generate_samplespecificabundance(mypath, phylo_name, nnum, thresholdi)
}
\arguments{
\item{mypath}{A character string specifying the path to the directory containing the data files.}

\item{phylo_name}{A character string specifying the name of the phyloseq RData file to load.}

\item{nnum}{An integer specifying the number of top taxa to select.}

\item{thresholdi}{A numeric value between 0 and 1 specifying the threshold for missing data.}
}
\value{
No return value. The function writes output files to the specified directory.
}
\description{
This function generates sample-specific abundance data by performing leave-one-out cross-validation (LOOCV) on phyloseq objects. It subsets the data based on specified parameters, fills missing values, and writes the output to files.
}
