% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_SSN.R
\name{generate_SSN}
\alias{generate_SSN}
\title{Generate SSN}
\usage{
generate_SSN(
  mypath = "E:/diabetes0409/",
  phylo_list = c("aphy_diabetesmetross_resp.Rdata")
)
}
\arguments{
\item{mypath}{The path to the directory containing the data files.}

\item{phylo_list}{A list of .Rdata files to be processed.}
}
\description{
This function generates SSN from the provided data.
}
